Fixed `config` symlink.
This config directory wasn't setup to properly symlink before, it should
now. As a part of fixing this, I brought over a bunch of stuff that was
already in my existing `config` dir. This also required adding some
`.gitignore` stuff to not include secretes from `gcloud` or load up on
default `node_modules`.