(vim) Updates teting to use `neovim`.
In practice I found I usually just wanted to have the built-in terminal
pop up during test runs.

I may want to revisit the set up and have the ability to run `neomake`
becuase the quickfix window is nice. The biggest issue I had is that a
test suite I work with has an exit code of 1 due to something in the
config - this makes the neomake function always think the tests failed.