(zsh) Remove light/dark mode toggling
Running the light/dark mode script was the slowest part of my zsh start
up time

num  calls                time                       self            name
-----------------------------------------------------------------------------------
 1)    1         134.38   134.38   74.33%    134.38   134.38   74.33%  light-mode
 2)    2          24.75    12.37   13.69%     24.75    12.37   13.69%  compaudit
 3)    1          39.82    39.82   22.02%     15.07    15.07    8.34%  compinit
 4)    2           2.66     1.33    1.47%      2.66     1.33    1.47%  _tags
 5)    1           4.79     4.79    2.65%      1.78     1.78    0.99%  _files
 6)    1           1.12     1.12    0.62%      1.12     1.12    0.62%  colors
 7)    1           5.27     5.27    2.91%      0.47     0.47    0.26%  _default
 8)    1           0.35     0.35    0.19%      0.35     0.35    0.19%  _have_glob_qual
 9)    2           0.13     0.07    0.07%      0.12     0.06    0.07%  ___git_complete
10)    1           5.34     5.34    2.95%      0.07     0.07    0.04%  _git
11)    2           0.01     0.01    0.01%      0.01     0.01    0.01%  complete

Since I am not using Kitty as much the need to set the theme is less
necessary. I suspect the slowest part was telling the OS which mode to
be in. If I want to re-enable theme switching, I may want separate out
the functionality to tell macOS what to set the theme to.