(git) Add .projectile files to gitignore
`.projectile` files can be used to tell projectile to treat a directory
as though it were a project. I am not working on a mono-repo project at
work, and want to be able to treat the different sub-directories as
their own projects; this means I am adding `.projectile` files to them.