(emacs) Update projectile-sort-order
Use the recently-active sort order for projectile.

A major reason for trying this was because in Phoenix projects I've
found I often want the context module, but it usually shows up last,
after all of the sub-modules in the namespace. This has proven to
generally put things in an order that I am happy with.