(macos) Fix zprofile for non-silicon Mac
Homebrew is installed to a different path on Apple Silicon than
Intel-based Macs[^1]. Last time I set up a machine, it was on Apple
Silicon, so I set a path in `zprofile` to use the silicon-specific path.
This failed to work on non-silicon Macs, so I have updated the
path-setting to be conditional.

[^1]: https://github.com/Homebrew/brew/issues/9177