(emacs) Update elixir-related configuration
I was running into issues with formatting via elixir-ls. I tried
explicitly updating the version of elixir-ls managed by the LSP package,
but that didn't seme to help.

For now, I am going to rely on elixir-mode to to do the formatting, now
elixir-ls.

I left the code to upgrade the elixir-ls version.