(macos) Add logseq to Brewfile
Install Logseq as a part of Brew download. While the current plan is to
continue to use Emacs as the desktop reader of org files, this may be
necessary to get iCloud to realized I want the Logseq-specific folder.
It may also not be necessary. For now, I think it will be useful to have
the app installed to possibly reference and configure things, but I can
revisit this if it's a pain.