(macos) Add script for remapping caps lock
Add a script to remap caps lock to control and have it run as a part of
`zprofile`. This is a limited and annoying, but makes it more scriptable
than the GUI-driven version that I currently do. Using a LaunchAgent may
be more robust and should be considered if I run into issues with this.